misc: change image slice scale to 200% of style's specified
Want to circle back and find a better pattern for this as it enlarges the slice window (which doesn't clamp the preview window) and it unnecessarily upscales small images which bloats storage a bit. There's probably also something smarter to do with just saving "as large as possible" rather than strictly template size.