fix: de-scale slice preview window
This is a workaround, not a good solution. The slice doesn't really handle sizing well and doesn't set great maximum sizes on the elements within it. I _think_ that can be worked around by constraining the preview, but none of that is really properly context aware to the user's screen.